🐛 Fix updateUserAssociations

+++ src/composables/useUserAssociations.ts
@@ -55,23 +55,25 @@ export default function () {
             isInOffice: false
         }
     ]
-    
+
     function updateUserAssociations(editedByStaff: boolean) {
         const instance = editedByStaff ? userManagerStore : userStore
         const userId = instance.user?.id
 
         userAssociations.value.forEach(async function (association) {
             // If we need to delete the association
-            if (association.deleteAssociation) {
+            if (association.id && association.deleteAssociation) {
                 await deleteUserAssociation(userId, association.id)
                 // Reactively remove the deleted item from interface
-                if (editedByStaff) return
-                userAssociations.value.splice(userAssociations.value.findIndex(obj => obj.id === association.id), 1)
-                userStore.user?.associations.splice(userAssociations.value.findIndex(obj => obj.id === association.id), 1)
+                if (!editedByStaff) {
+                    userAssociations.value.splice(userAssociations.value.findIndex(obj => obj.id === association.id), 1)
+                    userStore.user?.associations.splice(userAssociations.value.findIndex(obj => obj.id === association.id), 1)
+                }
             }
             // If we need to update the association
             else {
                 // We search for the corresponding association in store
+
                 const storeAssociation = instance.userAssociations.find(obj => obj.association?.id === association.id)
                 // We set a boolean to track changes
                 let hasChanges = false
@@ -85,16 +87,17 @@ export default function () {
                 else if (storeAssociation?.isVicePresident && association.role !== 'isVicePresident') hasChanges = true
                 else if (association.role !== 'isMember') hasChanges = true
 
-                if (!hasChanges) return
-                const infosToPatch = {
-                    isPresident: association.role === 'isPresident',
-                    canBePresidentFrom: association.canBePresidentFrom ? association.canBePresidentFrom : null,
-                    canBePresidentTo: association.canBePresidentTo ? association.canBePresidentTo : null,
-                    isSecretary: association.role === 'isSecretary',
-                    isTreasurer: association.role === 'isTreasurer',
-                    isVicePresident: association.role === 'isVicePresident'
+                if (hasChanges && association.id) {
+                    const infosToPatch = {
+                        isPresident: association.role === 'isPresident',
+                        canBePresidentFrom: association.canBePresidentFrom ? association.canBePresidentFrom : null,
+                        canBePresidentTo: association.canBePresidentTo ? association.canBePresidentTo : null,
+                        isSecretary: association.role === 'isSecretary',
+                        isTreasurer: association.role === 'isTreasurer',
+                        isVicePresident: association.role === 'isVicePresident'
+                    }
+                    await patchUserAssociations(userId, association.id, infosToPatch)
                 }
-                await patchUserAssociations(userId, association.id, infosToPatch)
             }
         })
     }
